ECE1_1/图像隐写/requirements.txt
numpy~=1.23.2
Pillow~=9.2.0